fails

+++ src/game/world/networking/Request.java
@@ -3,12 +3,12 @@ package game.world.networking;
 import java.io.Serializable;
 
 public class Request implements Serializable {
-    public int id;
+    public long id;
     public int keycode;
     public float x;
     public float y;
     
-    public Request(int id, int keycode, float x, float y){
+    public Request(long id, int keycode, float x, float y){
         this.id = id;
         this.keycode = keycode;
         this.x = x;
